Move into ai-input directory
